Moved balancer and resolver
